+++ attributes/default.rb
@@ -85,6 +85,13 @@ default['openstack']['compute']['network']['service_type'] = 'nova'
 # plugins from openstack-network
 default['openstack']['compute']['network']['plugins'] = ['openvswitch']
 
+# rootwrap.conf
+default['openstack']['compute']['rootwrap']['filters_path'] = '/etc/nova/rootwrap.d,/usr/share/nova/rootwrap'
+default['openstack']['compute']['rootwrap']['exec_dirs'] = '/sbin,/usr/sbin,/bin,/usr/bin'
+default['openstack']['compute']['rootwrap']['use_syslog'] = false
+default['openstack']['compute']['rootwrap']['syslog_log_facility'] = 'syslog'
+default['openstack']['compute']['rootwrap']['syslog_log_level'] = 'ERROR'
+
 # Neutron options
 default['openstack']['compute']['network']['neutron']['network_api_class'] = 'nova.network.neutronv2.api.API'
 default['openstack']['compute']['network']['neutron']['auth_strategy'] = 'keystone'